[Pinterest] Allow for scope descriptions
This makes a small change to the code introduced in https://github.com/microsoft/typespec/pull/849 to allow for models to be used as OAuth scopes.
Without this check, an object like `{value: "foo", description: "bar"}` gets wrapped in an additional `value` key.
When the OpenAPI3 emitter tries to print out the `securitySchemes`, it sees `{value: {...}}` and produces a scope with value `[object Object]` and no description.

Here instead we check for an object and prevent the wrapping.