Introduce `node-modules` config to override node module resolution
This commit introduces a new config option where one or more paths can be provided that contain `node_modules` directories.
When searching for a module, the resolver will prefer a module found in those paths over the default resolution order.

This is useful if you have a development copy of `@typespec` modules and you want to use those without changing the `import` statements in TSP files.

The option can either be provided in `tspconfig.yaml` or with the CLI flag `--node-modules`.