Add all inherited model properties to JSON data
When using `typespecTypeToJson` with a model type, only the properties implemented directly on that model are included.

Now we use `walkPropertiesInherited` to include all inherited properties as well.

This is technically a breaking change; specs that provide models with inherited properties to decorators using `typeSpecToJson` will now see those properties added to the output. The decorators affected by this are:
- `@useAuth` (from `http`)
- `@extension` (both `json-schema` and `openapi`)
- `@info` (from `openapi`)