Implement tsp namespace for http-client-csharp (#5443)
Fixes https://github.com/microsoft/typespec/issues/5442
Fixes https://github.com/microsoft/typespec/issues/5471
Fixes https://github.com/microsoft/typespec/issues/5563
Fixes https://github.com/Azure/azure-sdk-for-net/issues/47670

This PR contains the following:
1. now we honor the namespace defined in typespec for both model and
clients (short for "namespace is namespace" feature)
2. <del>removes the `namespace` configuration from the emitter because
we no longer need it. Azure generator still needs this, we will add it
back in azure emitter (temporary until azure generator decides how to
deal with it)</del>
3. removes the `use-model-namespace` configuration from the emitter
because we no longer need it. Azure generator still needs this, we will
add it back in azure emitter (temporary until azure generator decides
how to deal with it)
4. reports diagnostic when there are conflicts between namespace
segments and subclient names.

Because this change changes the namespace of literally everything, there
are quite a few file changed. Majority of them are the generated files
changed for the namespace.

We also have a piece of logic that automatically renames the client
`Models` to `ModelsOps` in our emitter. We should not have it in the
emitter any way. This part is removed. Azure part needs it, therefore we
will add the same logic back in azure generator.

Update:
I have to add the configuration `namespace` back because
`autorest.csharp` is still strongly coupled with this configuration,
removing it causes massive changes on that side. Therefore I decided
that we keep it here for now, we still write its value to
configuration.json for backward compatibility, but MGC will not read it
nor use it.