Standalone exe (#5794)
fix #5800
Adds a standalone executable. The way it works is by downloading the
latest compiler and running that.
This is right now the easiest option as it quite a pain to bundle the
package managers.

It comes with 2 command line installer

- install.sh: install on linux/macos
- powershell.ps1 (install on windows)

This right now doesn't work because the latest compiler doesn't have a
dependency on npm lib and expect to find a local `npm` exe which might
not be available if you did only install TypeSpec. SO we need to wait
next release to document it


There is quite a few follow up to work on
1. Add docs (After next release)
2. Include the install script in the website/better url for one
liner(e.g. `curl -fsSL https://typespec.io/install.sh | bash `)
3. Investigate more if we can bundle the cli (without the compiler) and
get rid of this installation step.

---------

Co-authored-by: Christopher Radek <14189820+chrisradek@users.noreply.github.com>