Fix: Infinite loop in language server due to not caching indeterminate entities (#5940)
fix https://github.com/microsoft/typespec/issues/2964
The problem is everytime we created an indeterminate entity we created a
new object instead of caching it from the type. This meant everytime we
needed to reinstantiate the template it would create 2 separate
instances.
This caused an inifinite loop in the LSP as it would call getTypeForNode
for diagnostics

This should also solve some duplicate type issue we might have
encountered.