[http-server-javascript] Implement basic project scaffolding. (#5906)
Adds a new scaffolding command for http-server-js. To run it, a customer
calls `npx hsj-scaffold`. It is opinionated, but uses some settings from
tspconfig.yaml.

- Generates a server entrypoint using either express or Node depending
on whether tspconfig.yaml has `express: true` in the emitter options.
- Generates a controller for each interface as a class that implements
the business logic interface. For now, these controller methods simply
throw 501 Not Implemented.
- Creates a suitable tsconfig.json to compile the project.
- Modifies package.json with scripts to build and start the server.
- Adds vscode tasks/launch configurations for debugging.

Some changes to the generator itself were required:
- Added a new helper module, `helpers/http`, that defines an
`HttpResponder` as an object containing a function that can respond to
an HTTP request. If the business logic implementations throw an object
that is an HttpResponder, then it will be used to respond to the request
instead of the default `onInternalError` handler. This makes it easy for
any route controller handler to construct and reply with any error they
like without having to satisfy the type checker.