[hsj] Enable SwaggerUI scaffolding (#6006)
- Adds an optional dependency on `@typespec/openapi3`.
- Fixed a bug in the openapi3 type definitions that caused a build
failure when I depend on `openapi3`.
- If openapi3 is installed, `http-server-js` emit will generate a file
`src/generated/openapi3.ts` that exports the OpenAPI v3 document
representing the service as a plain JS object inline.
- When scaffolding the project, if the server model is Express and
typespec/openapi3 is installed, additional hooks are generated to add
swaggerUI to the project, importing and using the generated document.
- I generate the OpenAPIv3 document at emit time so that it is always up
to date with the generated router. Then the scaffolding just references
it, so if you recompile the generated code but don't rescaffold, the
swaggerUI and anything else that might in the future be based on the
OpenAPI will stay in sync with the router, rather than with the
scaffolding.
- As a small bonus, I added request/response logging with `morgan` to
the scaffolded server, again if the server model is Express.