Efv2 to main (#5996)
# Efv2 Integration  

This PR introduces the work on **Emitter Framework v2** from the feature
branch into main. The existing emitter framework functionality remains
unchanged for now, the new framework is added as a new package
@typespec/emitter-framework

### **Core Changes**  

- **Compiler Alloy Support**  
- The compiler now detects if `$onEmit` returns an **Alloy component**.
If so, it loads Alloy and calls `render`.
  - This enables emitters to use Alloy components.  

- **New Package: `@typespec/emitter-framework`**  
- Wraps raw Alloy components to accept **TypeSpec types as props** and
automatically generate language elements.
- Currently includes a **subexport for TypeScript (`/typescript`)**,
with plans to extend support to other languages.
- Example: `ts.Interface` can take a `Model` or `Interface` and directly
render a TypeScript interface with model properties or operation
signatures.

- **New Package: `@typespec/http-client`**  
- Provides **TypeKits** for querying the **type graph** in the context
of an **HTTP client**.
- The goal is to eventually **power TCGC** and host decorators that make
sense in this context.
- Remains **decoupled from Azure**, but we may later introduce an
extension package (`@azure-tools/http-client-azure`) for Azure-specific
logic and decorators.

###  **Follow Up Work**  
- Deprecate EFv1
- A follow up API review and clean-up will be needed, specifically
around TypeKits