Some refactors around the CSharpEmitterContext and Logger (#5953)
Fixes #5845 

This PR is only refactoring. No functional changes.
1. Rename `NetEmitterOptions` and `NetEmitterOptionsSchema` to
`CSharpXXX` (resolves #5845 )
2. Move the type cache into `CSharpEmitterContext` to save a parameter
on functions.
3. Rearrange the order of parameters, making sure `CSharpEmitterContext`
is always the first.
4. Add `reportDiagnostic` method on `Logger`. Now we could use
`sdkContext.logger.reportDiagnostic` to report diagnostic. Theoretically
this method should have exact same parameter list as the original
`reportDiagnostics` except for the first. But since its type is quite a
complicated type and maybe in some circumstances, it might not resolve
entirely correctly.
5. Remove quite a few dead code.

---------

Co-authored-by: Weidong Xu <weidxu@microsoft.com>