[openapi3/json-schema] add `seal-object-schemas` emitter option (#5994)
This adds an emitter option to the Open API 3 and Json Schema emitters
to automatically set `additionalProperties: { not: {} }` and
`unevaluatedProperties: { not: {} }` for object schemas.

- This doesn't ovewrite the existing field if the field would already be
present (e.g. the field is already set to a different schema)
- The only applies to lead models. This means that when you have `model
A extends B`, `A` would automatically have this field set, but `B` would
not. This is because doing so would in most cases break validation for
the derived model.

I'm open to changing the name of the option if we want to make it
clearer that this only applies to leaf models. Since there are tradeoffs
with sealing base models there might be value in adding a future
`seal-object-schemas-strategy` option that defaults to `leaves-only` but
lets us support things like 'copy-base-models'

---------

Co-authored-by: Christopher Radek <Christopher.Radek@microsoft.com>