http-client-java, bug fix, parameter name mismatch in op and in convenienceApi (#6162)
`maxpagesize` get renamed to client name `maxPageSize`, but param in
convenienceApi was not renamed. This causes NPE when matching the
parameters.

Also add a protection in example generation code (with a warning), so
even in case there is problem, it won't crash there.

PR include some update to diagnostic, that now it includes all
stack/stdout/stderr, as I found it easier to get all the information
without re-run the whole process -- which may not be controlled by us.

autorest.java https://github.com/Azure/autorest.java/pull/3032