updates openapi `@extension` to support value types, and removes starts with `x-` extension name constraint (#6078)
Resolves #6076 
 
This PR makes a few changes to the `@typespec/openapi` `@extension`
decorator to align it closer to the `@typespec/json-schema` `@extension`
decorator.

`@extension` is updated in 3 ways:
1. The `x-` prefix is no longer required for the extension name.
2. Value kinds are now supported and will be inserted in the Open API
doc as raw values.
3. Passing in types (e.g. `typeof "foo"`, model expressions, tuple
types) retains current behavior but also triggers a deprecation warning.

## What's with the deprecation warning?
Prior to this PR, passed in types would be converted to JSON values (or
emit an error if they couldn't be converted).
So the following TypeSpec
```tsp
@OpenAPI.extension("x-foo", { foo: true })
model Foo {}
```
would emit the following OpenAPI:
```yaml
Foo:
  type: object
  x-foo:
    foo: true
```
However, in a follow-up release, passing in the type `{ foo: true }`
will cause an OpenAPI schema to be emitted instead:

```yaml
Foo:
  type: object
  x-foo:
    type: object
    required:
      - foo
    properties:
      foo:
        type: boolean
        enum:
          - true
```
To get back to the current behavior, `#{ foo: true }` should be passed
instead.

The deprecation warning gives users at least 1 release to be aware of
the change and update their code accordingly.

## Follow-ups

- typespec-azure PR: https://github.com/Azure/typespec-azure/pull/2236
- Update azure-specs repo to use values:
https://github.com/Azure/azure-rest-api-specs/pull/32832
- Update azure-specs-pr repo:
https://github.com/Azure/azure-rest-api-specs-pr/pull/21556

---------

Co-authored-by: Christopher Radek <Christopher.Radek@microsoft.com>