rest - filter string constant path http properties when using `@autoRoute` (#6326)
#6311 switched openapi3 to use `HttpProperty` instead of
`HttpOperationParameter`. The `@typespec/rest` package did some
filtering of parameters under a couple conditions (including path
constants) that wasn't applied to properties. This caused an extra path
parameter to be emitted in OpenAPI3 (and typespec-autorest when
eventually applying the same changes there).

This PR applies the same `@autoRoute` parameter filtering to properties.

---------

Co-authored-by: Christopher Radek <Christopher.Radek@microsoft.com>