Expose initScaffolding via an internals export (#6333)
In theory a better approach would probably move the init logic to a new
package but I think this would require quite a refactor of our compiler
to extract a lot of the utils.

Refactor a little bit to split compiler host from just a filesystem host
which non compile operation should depend on.

So this is a cheap solution that is only meant for our own use case for
now.