[core/http] Media type hinting and default content type changes for scalars. (#6357)
- Adds a decorator `@mediaTypeHint` that applies a hint to a Model,
Scalar, Union, or Enum that hints at how an emitter should serialize a
type by default. The hint is not itself mandatory, but libraries or
emitters may choose to honor it.
- Changed the logic for HTTP content type resolution.
- When resolving the content type of a payload's body, if the payload
has a content-type header, that type is used. Then, the type is resolved
to its lowest encoded type. If that type has a media type hint, the hint
type is used (this makes the media type hint mandatory by prescription
through the HTTP layer). If the type has no hint, then the default
content type for scalars is "text/plain" and for everything else is
"application/json".
- For multipart payloads, the default content type is "multipart/mixed"
for tuple-form multipart bodies and "multipart/form-data" for model-form
multipart bodies.
- The body of individual parts follows the same logic as non-multipart
payload bodies.

This changes the serialization of all scalars unless the content-type is
explicitly specified.

Checklist:
- [x] Fix remaining test that fails and figure out what to do with
`<scalar> | null`.
- [x] Fix failing tests in typespec-azure that result from these
changes.
- [x] Identify any specs in azure-rest-api-specs that need to have
content types explicitly specified.