[http] Fix handling of media types for literal types. (#6513)
fix #6511
This PR fixes the way that literal types are handled in media type
resolution. It treats them as equivalent to their closest scalars:

- String: TypeSpec.string
- StringLiteral: TypeSpec.string
- Boolean: TypeSpec.boolean
- Numeric: TypeSpec.numeric

Adds a regression test to ensure that in the case of an "extensible
enum" union, having both `string` and string literal variants results in
a single content-type: `"text/plain"`.