http-server-js - adds initial boilerplate for running http-specs (#6545)
This PR adds support for running the http-specs against the
http-server-js generated output.

Note: Currently `tsp-spector` doesn't collect coverage data for server
tests in a useful way so this data is not being uploaded.

The general flow is:
- Generate http-server-js output for all the http-specs. Specs that
cause http-server-js to crash are included in the `.testignore` file.
- Run vitest on all the e2e test files
- Each test file contains a single test that covers all the scenarios in
a single http spec file. 2 reasons:
1. This lets us trigger the mock api based on the spec file path
(roughly) instead of having to specify hard-coded scenario names (that
could in theory change)
2. The server output requires all routes to be defined, so it's easier
to do it all in one test rather than create individual tests with no-op
routes.
 

Test will show up in the vscode test extension if you open up the
http-server-js package in vscode (otherwise vitest chooses the workplace
config if you're in the root of the repo).

---------

Co-authored-by: Christopher Radek <Christopher.Radek@microsoft.com>