remove type conversion from openapi $extension (#6651)
Fixes https://github.com/microsoft/typespec/issues/6610

This PR disallows passing `Type` into the `@typespec/openapi`
`$extension` decorator function. This was meant to be disallowed when
switching to value syntax, and passing in `Type` to `@extension` will
eventually have new behavior:
https://github.com/microsoft/typespec/issues/6077

I added a diagnostic - this does not impact the codefix diagnostic when
using the `@extension` decorator from TypeSpec. Without the diagnostic,
the compiler hard crashes if passing in a `Type` to `$extension`. This
is a temporary measure until #6077 is resolved.

---------

Co-authored-by: Christopher Radek <Christopher.Radek@microsoft.com>