Align the client type shape from TCGC in our emitter  (#6179)
Fixes https://github.com/microsoft/typespec/issues/5924
Going forward to the ultimate goal of "only keep minimum code in our
emitter, which should only parse and spit out the data from TCGC", this
PR removes a few unnecessary code in our emitter, and align the
structure of client in our emitter with the client from TCGC.

Client initialization stuff has not been started therefore I still keep
the property `parameters` on the client type to keep maximum
compatibility and minimum the code change. I think the `parameters`
could be gone once our client initialization work item was done.

Operation stuff is also not changed, because the way TCGC represent
operations/methods is different from ours. I also keep that part as a
future work to align.

About client names:
Before, we did client name calculations in the emitter, to assign
verbose name to subclients, such as:
```
namespace Service;

namespace Foo {
	op one(): void;
	
	namespace Bar {
		op two(): void;
		
		namespace Qux {
			op three(): void;
		}
	}
}
```
We will have 4 clients. The hierarchy looks like this:
```
ServiceClient
- Foo
	- Bar
		- Qux
```
As a subclient, for instance, their full name could be
`ServiceClient.Foo.Bar` in this case.
Those second level (or deeper) subclients have a chance to have name
collisions because we would put them in the same namespace previously,
and we could have another subclient like `ServiceClient.Foo2.Bar`.

Therefore previously we are making name changes to them automatically
for those 2nd-level or deeper subclients by prepending all their
parents' name to it.

Now in this PR, it is not the goal here that we change that rule - this
PR just wants them all to unchange - therefore I put this logic at the
place that we have serialized the whole namespace to keep things in
parity.