[http] Implement bodyKind: file (#6563)
This PR implements separation of file bodies into their own bodyKind as
described in the design proposal.

- Adds a new bodyKind: "file"
- "file" bodyKind is returned when the effective type of the
request/response/part payload is-or-extends File, or if the explicit
body property (`@body`/`@bodyRoot`) has an effective type that
is-or-extends File.
- Extended `File` to be parameterized by the type of `contents`.
- Validates that the argument given to the `File` template is a scalar
and not a literal or union type that may be assignable to `string |
bytes`.
- Thoroughly tested file body scenarios.
- Preemptively fixed OpenAPI3 scenarios.

This addresses numerous bugs and unexpected behaviors around files and
solidifies the interpretation of files and their internal media types by
TypeSpec emitters.

Checklist:
- [x] Add tests to openapi3 validating the new scenarios work as
expected.
- [x] Validate that Http.File is not improperly extended by putting
invalid HTTP metadata on its properties.
- [x] Fix errors that result in type spec-azure.
- [x] Improve documentation of file bodies in handwritten and generated
docs for spec authoring and emitter direction.
- [x] Correct error targeting for case where `Contents` argument of
Http.File is a literal type.
- ~~[ ] Fix case where the `File` is in a union with other types.~~

---------

Co-authored-by: Christopher Radek <Christopher.Radek@microsoft.com>
Co-authored-by: Mark Cowlishaw <markcowl@microsoft.com>