[hsjs] Support scalar content-types and extended shared route differentiation. (#6898)
This PR does three things:

- Enables fallback content-type handling for types that are or extend
`TypeSpec.bytes`. Any unrecognized content-type is now allowed if the
body is-or-extends `bytes` and is deserialized as a Buffer.
- Enables text/plain serialization for types that are or extend
`TypeSpec.string`. This isn't enough to cover all scalars, but will
unblock `string` and types that extend it when the content type is
`text/plain`. In the future we should apply string decoding logic for
all scalars and treat this as equivalent to JSON for models.
- Enhances shared route differentiation. It now uses
`differentiateModelTypes` which is the same function used to split model
types apart in unions for response differentiation. The differentiation
logic was enhanced with a property filter, and for the purposes of
shared route differentiation I only allow headers. In the future this
could be extended to allow query parameters (by definition, there are no
different path parameters in a shared route).

This unblocks the following spector tests from compiling:

- encode/bytes
- payload/content-negotiation
- payload/media-type
- 
Closes https://github.com/microsoft/typespec/issues/3782
Closes https://github.com/microsoft/typespec/issues/6378
Closes https://github.com/microsoft/typespec/issues/6377