[http] Slightly stronger types for multipart resolution. (#6949)
This PR makes the types returned from `@typespec/http` for multipart
payloads slightly more accurate by splitting the tuple and model forms
of multipart bodies into separate types with their own `multipartKind`
and stricter subtypes about what kinds of parts they can have inside
them, whether or not they can be named, whether or not they are
optional, etc.

The types of the parts are also split into two `partKind` types: `tuple`
and `model`. The guarantees given are as follows:

**For model-based multipart payloads**:
- `type` is guaranteed to be a `Model`.
- `parts` are guaranteed to be only model-based parts.

**For tuple-based multipart payloads**:
- `type` is guaranteed to be a `Tuple`.
- `parts` are guaranteed to be only tuple-based parts.

**For model-based parts**:
- `name` is required.
- `property` is required and refers to the ModelProperty that defined
the _part_ (and we remember to set it -- we weren't doing this
correctly).

**For tuple-based parts**:
- `name` is optional.
- `property` is _guaranteed_ to be `undefined`.