[emitter-framework] Add support for ValueExpression (#6972)
This pull request introduces a new feature to add support for rendering
a Value Expression in the `@typespec/emitter-framework` package. The key
changes include the implementation of the `ValueExpression` component,
the addition of comprehensive tests for various value types, and the
initialization of an empty program for testing purposes.

### New Feature Implementation:
*
[`packages/emitter-framework/src/typescript/components/value-expression.tsx`](diffhunk://#diff-c767f5310b61b469266b95846986ad0b61ac675b975d886af8292a804243cf75R1-R56):
Implemented the `ValueExpression` component, which converts a TypeSpec
value into a JavaScript expression.

### Testing Enhancements:
*
[`packages/emitter-framework/test/typescript/components/value-expression.test.tsx`](diffhunk://#diff-df99d4d0d5d18ddbfb74cda34ed637f99c57829ef71d718ea822d7d3f110e355R1-R236):
Added tests to validate the rendering of various value types, including
strings, numbers, booleans, nulls, arrays, scalars, objects, and enums.
*
[`packages/emitter-framework/test/typescript/test-host.ts`](diffhunk://#diff-25c30ece2977ab4231fab89b603ad2566d5cfa50d83e78996e04da284f30c40eR42-R49):
Introduced the `initEmptyProgram` function to initialize an empty
program for testing purposes.