[python] Perf improvements (#6979)
The black plugin was causing the python emitter to take over 2-4 minutes
to run in some cases. For example, when using the python emitter with
tsp-client folks frequently run into this issue. Removed the black
plugin and moved that logic directly into the emitter. With these
updates the emitter takes about 15 seconds to run. ~90% improvement in
perf which will be a noticeable improvement for python emitter users.

---------

Co-authored-by: catalinaperalta <caperal@microsoft.com>