Fix telemetryClient initialization crash issue (#7021)
Fix https://github.com/microsoft/typespec/issues/7020

After adopt esbuild to build the library, the rollup changed, it will
crash when use `TelemetryReporter` as constructor to initialize
telemetry client. So update to use `TelemetryReporter.default` to refer
to the `TelemetryReporter` class and construct the instance. And apply
try-catch so that it will not crash the whole process when exception
occur during telemetry client initialization.