Remove the magic program from Typekits (#7018)
Fixes https://github.com/microsoft/typespec/issues/6789
This builds off of the POC PR
https://github.com/microsoft/typespec/pull/6865 from @joheredi

In short, this PR gets rid of the default typekit (e.g. being able to
use `$.api` without passing in a `program`). Now it is required to first
pass in a TypeSpec `program` or `realm` before calling Typekits (e.g.
`$(program).api`)

To reduce friction for emitter authors, the `emitter-framework` package
exposes the `useTsp()` hook which returns a typekit that's been
instantiated with a `program`. Also included an `Output` component that
accepts a `program` so that the new context provider doesn't need to be
manually created.

ToDo:
- [ ] compare performance of http-client-js with these changes compared
to using the global Default Typekit
- [ ] check impact on typespec-azure
https://github.com/Azure/typespec-azure/pull/2550

---------

Co-authored-by: Christopher Radek <Christopher.Radek@microsoft.com>