support nesting typekits (#7108)
Fixes https://github.com/microsoft/typespec/issues/7094

The caveat to this change is that you shouldn't have typekits that use
regular fields.

For example - `$.foo.type` <-- if this is a regular field that contains
an object, then that object will be wrapped in a Proxy, breaking object
identity.

However, we follow the pattern of using `get` descriptors for this
use-case instead, which would not break object identity.

---------

Co-authored-by: Christopher Radek <Christopher.Radek@microsoft.com>