Template parameter experimental (#7155)
Clarify that we don't want to make interactivng with template
declaration in the type graph a 1.0 feature.

The only library that did make use of that was tspd for doc generation
and is most likely the only use case that we should see today.

There is a few issues with teh template parameter type today
- No values
- No member acessing

To be able to find a correct solution to the above in the future we
keeping this as an expiermental feature