Fix the .tsp for the Spector test to specify base64url encoding (#7163)
Without that tile, the emitter returns `"base64"` as a value of
`SdkType.encode`, NOT `"base64url"`.

If the intent was to use the default, then maybe the `op base64()` on
the line 348 shouldn't be specifying `@encode`, but not this one.

Or, the tcgc should have a different default value, but from what I
currently see, the default appears to be `"base64"`, not `"base64url"`.

Either way, this fixes the spec, and now we can generate the correct
code to pass the test.

You may have a question - if this value has been incorrect all along,
how come that C# and Java emitters pass this test? (but C++, Go and
Python are still failing). And I think that is because Java and C# use a
function that does not complain when you ask it to decode base64 value,
but give it base64url, i.e. it can do both. In C++, the function that we
use is going to throw if you give it base64-encoded and not
base64-encoded value, so in our case the test was throwing an exception.

---------

Co-authored-by: Anton Kolesnyk <antkmsft@users.noreply.github.com>