typekits - add `$.type.resolve`, `$.value.resolve`, and update `$.resolve` to support returning values (#7167)
The `$.type.resolve` and `$.value.resolve` versions of the typekits also
support passing in a `kind` and will assert that the resolved type or
value matches the passed in kind.

Note: `$.resolve` wasn't actually available in the previous release, so
changing the return value to allow returning `Value` is not a breaking
change.

---------

Co-authored-by: Christopher Radek <Christopher.Radek@microsoft.com>