fix the empty client issue (#7160)
There are some RPs that are "models only RP", which contains no
operations but only models.
Such as we have in azure: Azure.Messaging.EventGrid.SystemEvents.

There was an update in TCGC 0.54.1, now no client is returned from
TCGC's output.
And long time ago, we made an update in our emitter to error out when
there is no client in TCGC's output.
When these two combined, we get errors when generating
Azure.Messaging.EventGrid.SystemEvents.

I think it is not quite reasonable that we give the user an error when
there is no client - a generator should work fine if they just want to
generate the models.
Therefore this PR is changing that to a warning, and we still return a
reasonable code model when there is no client.