change $.model.getSpreadType to $.model.getIndexType (#7202)
What `getSpreadType` returned is actually the `model.indexer.value`
wrapped in a `Record` or `Array` model, depending on what the
`model.indexer.key` is. This is different than whatever is spread.

2 examples of how this is different:
- Multiple records: `model Foo { ...Record<string>; ...Record<int8>; }`
has an indexer value that is the union of `string | int8`.
- Excludes non-indexed types: `model Foo { ...Bar }` would not return
the `Bar` type as expected by the typekit name.

## Just remove getSpreadType instead of rename it?
I considered just removing getSpreadType but there are a few benefits to
having `getIndexType`
- No need to check the `model.indexer.key.name` to determine if the
indexer was sourced from an `Array` or `Record` type.
- `getIndexType` supports caching, so the `Record`/`Array` type is only
created once per model when needed, instead of every time it's
requested. Given types can be used as refkeys in alloy this seems like a
useful feature.

---------

Co-authored-by: Christopher Radek <Christopher.Radek@microsoft.com>