[core] Support renaming and caching in visibility transforms. (#7200)
This PR adds some features and fixes a couple of bugs in the visibility
transforms.

- `@withVisibilityFilter` now accepts a `nameTemplate` optional argument
that is used to rename named types as they are transformed.
- `@withVisibilityFilter` now correctly _completely ignores_ any type
that is not subject to visibility transforms. Previously, it was still
mutating/cloning types that were not subject to the transform, like
Scalar types, even if it didn't transform them.
- `@withVisibilityFilter` now caches transforms so that applications of
the same visibility filter with the same name template now return
identical result objects, reducing the number of structurally equivalent
models/unions that are created.
- `@withVisibilityFilter` now correctly considers the indexer of Model
types, allowing it to work correctly on Array and Record types.
- `@withLifecycleUpdate` also receives the above modifications.

Confirmed that this addresses the following issues in a local playground
instance.

Closes #7016 
Closes #7072 

TODO before merging:
- [x] Add tests for regressions based on the linked issues.