http-client-java, prepare release 0.4.0 (#7267)
`isAzure` is very confusing, as we now have `isAzureV1`, `isAzureV2`
(that I thought this `isAzure` means "either v1 or v2", but it is not).
It should not have effect in track2. We always use `--fluent` hence
`isFluent` or `isAzureOrFluent` (this get removed too).

merged with https://github.com/microsoft/typespec/pull/7268

add adaption to tcgc 0.56.1, as it affects behavior

---------

Co-authored-by: XiaofeiCao <xiaofeicao@microsoft.com>