[hsjs] First pass of canonical visibility transforms. (#7256)
This is part one of Visibility for HSJS.

- In core, exposes `VisibilityFilter.toCacheKey` so that HSJS can use
the same logic as core visibility to cache mutator results by visibility
filter without worrying about object instance identity.
- Adds a `canonicalizeHttpOperation` function that accepts an
`Operation` and returns a canonically transformed version of the
operation.
- The canonical operation has visibility applied per the rules of
`@typespec/http`.
- The canonical operation has implicit optionality applied using
`MetadataInfo`.
- The canonical operation recognizes cases where we do not understand
the encoding described by `@encode` and converts such properties/scalars
into their target types.
- Across the emitter, anywhere we analyze operation signatures, we first
canonicalize the operation. If necessary, we also recompute the
`HttpOperation` for the canonical operation. This allows existing
"visibility naive" code to continue to work without many modifications,
as the canonical transform simply applies the necessary changes to a new
copy of the operation.

This **does not** fix some existing issues with deeply-nested metadata.
Fixes for that will come as part of work to address
https://github.com/microsoft/typespec/issues/6934

Closes https://github.com/microsoft/typespec/issues/6376