[hsjs] Don't consider node built-in modules external dependencies (#7280)
The scaffolding script incorrectly considers node built-in modules to be
external dependencies. This can cause the scaffolding script to fail in
some cases because it cannot find a dependency version to use for those
builtins. The fix is to simply exclude node builtins like `node:http`
from the external dependencies in the same way we currently exclude
absolute and relative paths.